+++ packages/go_router/lib/src/route.dart
@@ -13,6 +13,7 @@ import 'package:meta/meta.dart' as meta;
 
 import 'configuration.dart';
 import 'match.dart';
+import 'misc/merged_observer.dart';
 import 'path_utils.dart';
 import 'router.dart';
 import 'state.dart';
@@ -498,7 +499,19 @@ abstract class ShellRouteBase extends RouteBase {
     super.redirect,
     required super.routes,
     required super.parentNavigatorKey,
+    this.mergeObservers = true,
   }) : super._();
+
+  /// Whether to merge the observers of the shell route's parent with the
+  /// observers of the shell route.
+  ///
+  /// When `true`, the observers of the shell route's parent will be merged with
+  /// the observers of the shell route.
+  ///
+  /// Only effective for `observers` passed into `GoRouter`.
+  ///
+  /// Defaults to `false`.
+  final bool mergeObservers;
 
   static void _debugCheckSubRouteParentNavigatorKeys(
     List<RouteBase> subRoutes,
@@ -577,14 +590,30 @@ class ShellRouteContext {
   final NavigatorBuilder navigatorBuilder;
 
   Widget _buildNavigatorForCurrentRoute(
+    BuildContext context,
     List<NavigatorObserver>? observers,
+    bool mergeObservers,
     String? restorationScopeId,
   ) {
+    final List<NavigatorObserver> effectiveObservers = <NavigatorObserver>[
+      ...?observers
+    ];
+
+    if (mergeObservers) {
+      final List<NavigatorObserver>? rootObservers =
+          GoRouter.maybeOf(context)?.observers;
+      if (rootObservers != null) {
+        effectiveObservers.add(MergedNavigatorObserver(
+          rootObservers,
+        ));
+      }
+    }
+
     return navigatorBuilder(
       navigatorKey,
       match,
       routeMatchList,
-      observers,
+      effectiveObservers,
       restorationScopeId,
     );
   }
@@ -691,6 +720,7 @@ class ShellRoute extends ShellRouteBase {
     super.redirect,
     this.builder,
     this.pageBuilder,
+    super.mergeObservers,
     this.observers,
     required super.routes,
     super.parentNavigatorKey,
@@ -732,7 +762,9 @@ class ShellRoute extends ShellRouteBase {
   ) {
     if (builder != null) {
       final Widget navigator = shellRouteContext._buildNavigatorForCurrentRoute(
+        context,
         observers,
+        mergeObservers,
         restorationScopeId,
       );
       return builder!(context, state, navigator);
@@ -748,7 +780,9 @@ class ShellRoute extends ShellRouteBase {
   ) {
     if (pageBuilder != null) {
       final Widget navigator = shellRouteContext._buildNavigatorForCurrentRoute(
+        context,
         observers,
+        mergeObservers,
         restorationScopeId,
       );
       return pageBuilder!(context, state, navigator);
@@ -870,6 +904,7 @@ class StatefulShellRoute extends ShellRouteBase {
     super.redirect,
     this.builder,
     this.pageBuilder,
+    super.mergeObservers,
     required this.navigatorContainerBuilder,
     super.parentNavigatorKey,
     this.restorationScopeId,
@@ -900,6 +935,7 @@ class StatefulShellRoute extends ShellRouteBase {
   /// for a complete runnable example using StatefulShellRoute.indexedStack.
   StatefulShellRoute.indexedStack({
     required List<StatefulShellBranch> branches,
+    bool mergeObservers = true,
     GoRouterRedirect? redirect,
     StatefulShellRouteBuilder? builder,
     GlobalKey<NavigatorState>? parentNavigatorKey,
@@ -911,6 +947,7 @@ class StatefulShellRoute extends ShellRouteBase {
          redirect: redirect,
          builder: builder,
          pageBuilder: pageBuilder,
+         mergeObservers: mergeObservers,
          parentNavigatorKey: parentNavigatorKey,
          restorationScopeId: restorationScopeId,
          navigatorContainerBuilder: _indexedStackContainerBuilder,
@@ -1018,6 +1055,7 @@ class StatefulShellRoute extends ShellRouteBase {
   ) => StatefulNavigationShell(
     shellRouteContext: shellRouteContext,
     router: GoRouter.of(context),
+    mergeObservers: mergeObservers,
     containerBuilder: navigatorContainerBuilder,
   );
 
@@ -1207,6 +1245,7 @@ class StatefulNavigationShell extends StatefulWidget {
     required this.shellRouteContext,
     required GoRouter router,
     required this.containerBuilder,
+    this.mergeObservers = true,
   }) : assert(shellRouteContext.route is StatefulShellRoute),
        _router = router,
        currentIndex = _indexOfBranchNavigatorKey(
@@ -1230,6 +1269,9 @@ class StatefulNavigationShell extends StatefulWidget {
   ///
   /// Corresponds to the index in the branches field of [StatefulShellRoute].
   final int currentIndex;
+
+  /// Same to [ShellRoute.mergeObservers].
+  final bool mergeObservers;
 
   /// The associated [StatefulShellRoute].
   StatefulShellRoute get route => shellRouteContext.route as StatefulShellRoute;
@@ -1422,7 +1464,9 @@ class StatefulNavigationShellState extends State<StatefulNavigationShell>
         previousBranchLocation != currentBranchLocation;
     if (locationChanged || !hasExistingNavigator) {
       branchState.navigator = shellRouteContext._buildNavigatorForCurrentRoute(
+        context,
         branch.observers,
+        false,
         branch.restorationScopeId,
       );
     }